Rag_chatbot/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
transformers==4.35.2
sentence-transformers==2.2.2
faiss-cpu==1.13.2
python-dotenv==1.0.0
pydantic==2.5.2
chainlit
requests
